feat(level-*.md): provide an English translation
Thanks to DeepL,[1] it was possible to machine translate the levels'
markdown files into English (US).  By copy-paste of the Chinese
content into the engine's interface, some of the commands and notes
are now present at least twice (i.e. the translation into English,
the notes in English by the author of the source strings).  This,
spell checks/edits of an eventual English version, and improvements
markdownlint will suggest will be subject of subsequent commits.

[1] https://www.deepl.com/en/translator, the free / non-subscription
based version available via the web broswer was used.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>